feat: offline search screen
