+++ android/app/src/main/AndroidManifest.xml
@@ -2,6 +2,7 @@
   package="com.timescomponents">
 
     <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 
     <application
       android:name=".MainApplication"
