+++ packages/search/__tests__/ios/cancel-button.test.tsx
@@ -1,5 +1,5 @@
 import TestRenderer from "react-test-renderer";
-import Magnifier from "@times-components-native/search/src/search-bar/magnifier";
+import Magnifier from "@times-components-native/search/src/searchBar/magnifier";
 import React from "react";
 
 describe("CancelButton", () => {
