+++ packages/search/__tests__/ios/x-button.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import TestRenderer from "react-test-renderer";
-import XButton from "../../src/search-bar/x-button";
+import XButton from "../../src/searchBar/xButton";
 
 describe("XButton", () => {
   it("Should render correctly", () => {
