+++ packages/search/__tests__/search-bar-shared.tsx
@@ -1,8 +1,8 @@
 import TestRenderer from "react-test-renderer";
 import React from "react";
-import EmptySearchMessage from "../src/empty-search-message";
-import CancelButton from "../src/search-bar/cancel-button";
-import { SearchBar } from "../src/search-bar/search-bar";
+import EmptySearchMessage from "../src/emptySearchMessage";
+import CancelButton from "../src/searchBar/cancelButton";
+import { SearchBar } from "../src/searchBar/searchBar";
 
 export default () => {
   it("EmptySearchMessage", () => {
