+++ packages/search/src/searchBar/styles/chevronStyles.ts
@@ -1,0 +1,10 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+  container: {
+    position: "absolute",
+    top: 6,
+    left: 0,
+    zIndex: 1,
+  },
+});
