+++ packages/search/src/searchBar/styles/magnifierStyles.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+  container: {
+    position: "absolute",
+    top: 6,
+    left: 4,
+  },
+});
