+++ packages/search/src/searchBar/styles/xButtonStyles.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+  closeIconContainer: {
+    position: "absolute",
+    right: 10,
+    top: 5,
+  },
+});
