+++ packages/search/src/searchBar/xButton.tsx
@@ -1,10 +1,7 @@
 import React, { FC } from "react";
-import {
-  StyleSheet,
-  TouchableOpacity,
-  TouchableOpacityProps,
-} from "react-native";
+import { TouchableOpacity, TouchableOpacityProps } from "react-native";
 import CloseRoundIcon from "@times-components-native/icons/src/icons/close-round";
+import { styles } from "./styles/xButtonStyles";
 
 export interface XButtonProps {
   onPress: TouchableOpacityProps["onPress"];
@@ -15,13 +12,5 @@ const XButton: FC<XButtonProps> = ({ onPress }) => (
     <CloseRoundIcon />
   </TouchableOpacity>
 );
-
-const styles = StyleSheet.create({
-  closeIconContainer: {
-    position: "absolute",
-    right: 10,
-    top: 5,
-  },
-});
 
 export default XButton;
