+++ packages/search/src/searchList/styles/searchListItemSnippetStyles.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from "react-native";
+import { colours } from "@times-components-native/styleguide";
+
+export const styles = StyleSheet.create({
+  text: {
+    color: colours.functional.secondary,
+  },
+  hightlighted: {
+    color: colours.functional.black,
+  },
+});
