+++ packages/search/src/searchList/styles/searchListLoaderStyles.ts
@@ -1,0 +1,8 @@
+import { StyleSheet } from "react-native";
+import { spacing } from "@times-components-native/styleguide";
+
+export const styles = StyleSheet.create({
+  loader: {
+    padding: spacing(4),
+  },
+});
