+++ packages/utils/src/platformUtils.ts
@@ -1,0 +1,3 @@
+import { Platform } from "react-native";
+
+export const isIOS = Platform.OS === "ios";
